Began the density_algorithm
I have started a new file that will handle the density algorithm. This
is where the creativity and collaboration wil come into play. We may
want to consider doing multiple branches at this point to try different
algorithms. The main goal of the density algorithm at this point is to
take in the output of the sentencer class and chop the sentences into
"bits of information". I have a really primitive algorithm that just
looks for nouns and commas basically. Be thinking of ideas.